requirements.txt
selenium
Pillow
webdriver_manager
yahoo_fin
functions